// ClientApp/src/app/models/docente.ts
export class Docente {

    id: number;
    nombre: string;
    segundonombre:string;
    apellido: string;
    segundoapellido:string;
    telefono: number;
    sexo: boolean;
    fechaNacimiento: string;
    emailInstitucional: string;
    emailPersonal: string;
    programa: string;
    CvLAC: string;
    estado: boolean;
    direccion: string;




}